Update to 'Revision 2' - HSI mode
This update changed a LOT of the underlying color logic. Instead of
trying to work things out in RGB space, it converts things to HSI which
allows some really easy looping of colors. A left a lot of commented out
code in this commit (for various trail/errors) that will get cleaned up
later.