Adding other testing sketches
Adding the 'ButtonTesting' sketch which was an attempt at trying to
figure out how to do a long-press; it seems to work fine on my proto
board, but not at all on the final PCB/wired one. This also adds the
'HsiColor' sketch which is where I ironed out all the bugs to HSI before
adding it to the main ArduinoRgbController.