Using IRremote 'isIdle' call for reliability
I found that using the 'isIdle' call while receiving a signal is
essential in order to 'listen' to the entire length of the IR signal.
This commit also makes a note that setting the 'TOLERANCE' variable
within IRremoteInt.h (line 130 ish) to around 40 instead of 25 also
helps with reliability -> changes to the library are not controlled
here.