Added check for holding an IR button down (repeat)
This checks whether the IR remote is sending an 0xFFFFFFFF which
seemingly represents a 'resend' or 'repeat' concept. The logic in this
commit checks for this value and then re-processes the last saved button
press again. This allows the user to hold the (e.g.) brightness button
down and have it repeatedly turn the brightness up/down.